refactored pending draft notifications to use AlarmManager
