+++ WordPress/src/main/AndroidManifest.xml
@@ -495,10 +495,6 @@
             android:name=".push.NotificationsScreenLockWatchService"
             android:exported="false"
             android:label="Notifications Screen Lock watch Service" />
-        <service
-            android:name=".ui.notifications.services.NotificationsPendingDraftsService"
-            android:exported="false"
-            android:label="Notifications Pending Drafts Service" />
 
         <!-- Samsung multiwindow support -->
         <uses-library
