+++ WordPress/src/main/java/org/wordpress/android/ui/posts/PostsListFragment.java
@@ -27,11 +27,11 @@ import org.wordpress.android.models.PostsListPostList;
 import org.wordpress.android.push.NativeNotificationsUtils;
 import org.wordpress.android.ui.ActivityLauncher;
 import org.wordpress.android.ui.EmptyViewMessageType;
+import org.wordpress.android.ui.notifications.receivers.NotificationsPendingDraftsReceiver;
 import org.wordpress.android.ui.posts.adapters.PostsListAdapter;
 import org.wordpress.android.ui.posts.services.PostEvents;
 import org.wordpress.android.ui.posts.services.PostUpdateService;
 import org.wordpress.android.ui.posts.services.PostUploadService;
-import org.wordpress.android.ui.prefs.AppPrefs;
 import org.wordpress.android.util.AniUtils;
 import org.wordpress.android.util.NetworkUtils;
 import org.wordpress.android.util.ToastUtils;
@@ -45,7 +45,7 @@ import org.xmlrpc.android.ApiHelper.ErrorType;
 
 import de.greenrobot.event.EventBus;
 
-import static org.wordpress.android.ui.notifications.services.NotificationsPendingDraftsService.PENDING_DRAFTS_NOTIFICATION_ID;
+import static org.wordpress.android.ui.notifications.receivers.NotificationsPendingDraftsReceiver.PENDING_DRAFTS_NOTIFICATION_ID;
 
 public class PostsListFragment extends Fragment
         implements PostsListAdapter.OnPostsLoadedListener,
@@ -461,9 +461,6 @@ public class PostsListFragment extends Fragment
         post.setPostStatus(PostStatus.toString(PostStatus.PUBLISHED));
         post.setChangedFromDraftToPublished(true);
 
-        // also in case this postId was in our ignore list, delete it from the list as well
-        AppPrefs.deleteIdFromAllPendingDraftsIgnorePostIdLists(post.getLocalTablePostId());
-
         PostUploadService.addPostToUpload(post);
         getActivity().startService(new Intent(getActivity(), PostUploadService.class));
 
@@ -543,16 +540,8 @@ public class PostsListFragment extends Fragment
                     mShouldCancelPendingDraftNotification = false;
 
                     // delete the pending draft notification if available
-                    NativeNotificationsUtils.dismissNotification(PENDING_DRAFTS_NOTIFICATION_ID, getActivity());
-
-                    // note that cancelling the notification dismisses not only the case where the notification was
-                    // about this very local draft but will also dismiss it even if there were several outstanding
-                    // pending drafts.
-                    // We don't re-run the service here to notify the user of other  pending drafts, because the
-                    // user is already looking at the blog post list, so it doesn't make sense bothering them
-
-                    // also in case this postId was in our ignore list, delete it from the list as well
-                    AppPrefs.deleteIdFromAllPendingDraftsIgnorePostIdLists(post.getPostId());
+                    int pushId = NotificationsPendingDraftsReceiver.makePendingDraftNotificationId(post.getPostId());
+                    NativeNotificationsUtils.dismissNotification(pushId, getActivity());
                 }
             }
         });
