+++ WordPress/src/main/java/org/wordpress/android/ui/prefs/AppPrefs.java
@@ -77,14 +77,6 @@ public class AppPrefs {
         // local time of the last push notification received
         PUSH_NOTIFICATIONS_LAST_NOTE_TIME,
 
-        // list for local drafts that the user has set to ignore notifications on
-        // (i.e. "please don't remind me I've got this in drafts")
-        PENDING_DRAFTS_NOTIFICATION_IGNORE_LIST_MONTH,
-
-        PENDING_DRAFTS_NOTIFICATION_IGNORE_LIST_WEEK,
-
-        PENDING_DRAFTS_NOTIFICATION_IGNORE_LIST_DAY,
-
         // local IDs of sites recently chosen in the site picker
         RECENTLY_PICKED_SITE_IDS,
     }
@@ -462,104 +454,6 @@ public class AppPrefs {
 
     public static void setReaderSwipeToNavigateShown(boolean alreadyShown) {
         setBoolean(UndeletablePrefKey.SWIPE_TO_NAVIGATE_READER, alreadyShown);
-    }
-
-    public static ArrayList<Long> getPendingDraftsIgnorePostIdList(DeletablePrefKey listToWorkOn) {
-        // builds the list from csv
-        ArrayList<Long> idList = new ArrayList<>();
-        String idListString = getString(listToWorkOn, "");
-        if (!TextUtils.isEmpty(idListString)) {
-            List<String> items = Arrays.asList(idListString.split(","));
-            for (String item : items) {
-                Long oneId = Long.valueOf(item);
-                idList.add(oneId);
-            }
-        }
-        return idList;
-    }
-
-    public static void addToPendingDraftsIgnorePostIdList(long postId, DeletablePrefKey listToWorkOn) {
-        // appends new id to postIdList, unless it's already in the list
-        // also rolls - eliminates first object (i..e oldest object) when MAX_PENDING_DRAFTS_AMOUNT reached
-        boolean isPostIdAlreadyInIgnoreList = isPostIdInPendingDraftsIgnorePostIdList(postId, listToWorkOn);
-
-        // if this postId wasn't in our list already, append it
-        if (!isPostIdAlreadyInIgnoreList) {
-
-            // before appending, eliminate first one if we have more than MAX_PENDING_DRAFTS_AMOUNT elements
-            rollPendingDraftsIgnorePostIdListIfNeeded(listToWorkOn);
-
-            String idListString = getString(listToWorkOn, "");
-            if (!TextUtils.isEmpty(idListString)) {
-                idListString += "," + String.valueOf(postId);
-            } else {
-                idListString += String.valueOf(postId);
-            }
-            setString(listToWorkOn, idListString);
-        }
-    }
-
-    public static void addToPendingDraftsIgnorePostIdList(ArrayList<Long> postIds, DeletablePrefKey listToWorkOn) {
-        if (postIds != null) {
-            for (Long postId : postIds) {
-                addToPendingDraftsIgnorePostIdList(postId, listToWorkOn);
-            }
-        }
-    }
-
-    public static void deleteIdFromAllPendingDraftsIgnorePostIdLists(long postId) {
-        deleteIdFromPendingDraftsIgnorePostIdList(postId, DeletablePrefKey.PENDING_DRAFTS_NOTIFICATION_IGNORE_LIST_DAY);
-        deleteIdFromPendingDraftsIgnorePostIdList(postId, DeletablePrefKey.PENDING_DRAFTS_NOTIFICATION_IGNORE_LIST_WEEK);
-        deleteIdFromPendingDraftsIgnorePostIdList(postId, DeletablePrefKey.PENDING_DRAFTS_NOTIFICATION_IGNORE_LIST_MONTH);
-    }
-
-    public static void deleteIdFromPendingDraftsIgnorePostIdList(long postId, DeletablePrefKey listToWorkOn) {
-        ArrayList<Long> idList = getPendingDraftsIgnorePostIdList(listToWorkOn);
-        boolean foundSamePostId = false;
-        for (Long oneId : idList) {
-            if (oneId != null && oneId == postId) {
-                // now delete this from the list and rebuild it
-                idList.remove(oneId);
-                foundSamePostId = true;
-                break;
-            }
-        }
-
-        if (foundSamePostId) {
-            // re-build
-            String idListString = TextUtils.join(",", idList);
-            setString(listToWorkOn, idListString);
-        }
-    }
-
-    private static boolean rollPendingDraftsIgnorePostIdListIfNeeded(DeletablePrefKey listToWorkOn){
-        String idListString = getString(listToWorkOn, "");
-        boolean rolled = false;
-        if (!TextUtils.isEmpty(idListString)) {
-            // note wrapping the Arrays.asList call with a new object is needed because otherwise
-            // trying to remove an item from a List returned by Arrays.asList throws an UnsupportedOperationException
-            List<String> items = new ArrayList(Arrays.asList(idListString.split(",")));
-            if (items.size() > MAX_PENDING_DRAFTS_AMOUNT) {
-                // eliminate first one
-                items.remove(0);
-                idListString = TextUtils.join(",", items);
-                setString(listToWorkOn, idListString);
-                rolled = true;
-            }
-        }
-        return rolled;
-    }
-
-    public static boolean isPostIdInPendingDraftsIgnorePostIdList(long postId, DeletablePrefKey listToWorkOn) {
-        ArrayList<Long> idList = getPendingDraftsIgnorePostIdList(listToWorkOn);
-        boolean foundSamePostId = false;
-        for (Long oneId : idList) {
-            if (oneId != null && oneId == postId) {
-                foundSamePostId = true;
-                break;
-            }
-        }
-        return foundSamePostId;
     }
 
     /*
